Some changes

+++ blog/api/views.py
@@ -1,6 +1,8 @@
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view ,permission_classes ,authentication_classes
+from rest_framework.permissions import IsAuthenticated
+
 
 from account.models import Account
 from blog.models import BlogPost
@@ -14,7 +16,7 @@ CREATE_SUCCESS = 'created'
 
 
 @api_view(['GET', ])
-# @permission_classes([])
+@permission_classes([IsAuthenticated])
 # @authentication_classes([])
 def api_detail_blog_view(request, blog_id):
 
@@ -22,6 +24,11 @@ def api_detail_blog_view(request, blog_id):
 		blog_post = BlogPost.objects.get(id = blog_id)
 	except BlogPost.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
+
+	#user = request.user
+	#print(user)
+	#print(blog_post.author)
+
 
 	if request.method == 'GET':
 		serializer = BlogPostSerializer(blog_post)
@@ -36,6 +43,13 @@ def api_update_blog_view(request, blog_id):
 	except BlogPost.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
 
+	user = request.user
+	#print(user)
+	#print(blog_post.author)
+
+	if blog_post.author != user:
+		return Response({"message" : "You can't update this post because you have not created this post"})
+
 	if request.method == 'PUT':
 		serializer = BlogPostSerializer(blog_post, data=request.data)
 		data = {}
@@ -47,6 +61,7 @@ def api_update_blog_view(request, blog_id):
 
 
 @api_view(['DELETE',])
+@permission_classes([IsAuthenticated])
 # @permission_classes([])
 # @authentication_classes([])
 def api_delete_blog_view(request, blog_id):
@@ -55,6 +70,14 @@ def api_delete_blog_view(request, blog_id):
 		blog_post = BlogPost.objects.get(id = blog_id)
 	except :
 		return Response(status=status.HTTP_404_NOT_FOUND)
+	
+	user = request.user
+	#print(user)
+	#print(blog_post.author)
+
+	if blog_post.author != user:
+		return Response({"message" : "You can't delete this post because you have not created this post"})
+
 
 	if request.method == 'DELETE':
 		operation = blog_post.delete()
@@ -65,11 +88,10 @@ def api_delete_blog_view(request, blog_id):
 
 
 @api_view(['POST'])
-@permission_classes([])
-@authentication_classes([])
+@permission_classes([IsAuthenticated])
 def api_create_blog_view(request):
 
-	account = Account.objects.get(pk=6)
+	account = request.user
 
 	blog_post = BlogPost(author=account)
